Add a few more renames
