+++ src/System.Windows.Forms/src/System/Windows/Forms/OLE/DataStore.DataStoreEntry.cs
@@ -1,7 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace System.Windows.Forms;
+namespace System.Private.Windows.Ole;
 
 internal sealed partial class DataStore
 {
